requirements.txt
firecrawl-py
python-dotenv
anthropic
